refactor:PostDetail 컴포넌트 분리
